+++ guildwars2api/__init__.py
@@ -1,2 +1,4 @@
 # -*- coding: utf-8 -*-
-__version__ = '0.6.0'
+__version__ = '0.7.0'
+
+from base import GuildWars2APIError